% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_calibration.R
\name{log_prior_age_calibration}
\alias{log_prior_age_calibration}
\title{Log-prior for the joint two-acid age calibration}
\usage{
log_prior_age_calibration(
  log_a_Asp,
  e_Asp,
  log_sigma_Asp,
  log_a_Glu,
  e_Glu,
  log_sigma_Glu,
  z_rho
)
}
\arguments{
\item{log_a_Asp, e_Asp, log_sigma_Asp}{TDK (Asp) calibration parameters:
log scale factor, exponent, log residual SD (log-age scale).}

\item{log_a_Glu, e_Glu, log_sigma_Glu}{SPK (Glu) calibration parameters.}

\item{z_rho}{Fisher-z transform of the residual correlation
(\eqn{\rho = \tanh(z_\rho)}).}
}
\value{
Scalar log-prior.
}
\description{
Weakly informative priors centred on literature-scale values (Kaufman
et al. 2013-type Np calibrations), wide enough that a calibration
dataset of ~100+ samples dominates the posterior.
}
