% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_calibration_forward.R
\name{plot_age_calibration_fit_fwd}
\alias{plot_age_calibration_fit_fwd}
\title{Plot forward-model predicted vs. observed D/L for the calibration set}
\usage{
plot_age_calibration_fit_fwd(
  mcmc_out,
  precomputed,
  burnin = 1000,
  dT_yr = 1500
)
}
\arguments{
\item{mcmc_out}{List returned by \code{\link{run_mcmc}} using
\code{\link{log_posterior_age_calibration_fwd}}.}

\item{precomputed}{List returned by \code{\link{precompute_forward_inputs}}.}

\item{burnin}{Integer. Burn-in iterations to discard. Default \code{1000}.}

\item{dT_yr}{Integration timestep (yr). Default \code{1500}.}
}
\value{
A ggplot object.
}
\description{
Plot forward-model predicted vs. observed D/L for the calibration set
}
