% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_calibration_forward.R
\name{summarize_calibration_fwd}
\alias{summarize_calibration_fwd}
\title{Extract posterior-mean forward-model calibration parameters}
\usage{
summarize_calibration_fwd(mcmc_out, burnin = 1000)
}
\arguments{
\item{mcmc_out}{List returned by \code{\link{run_mcmc}} using
\code{\link{log_posterior_age_calibration_fwd}}.}

\item{burnin}{Integer. Burn-in iterations to discard. Default \code{1000}.}
}
\value{
A named list with elements \code{Ae_Asp}, \code{Ea_Asp}, \code{x_Asp},
\code{sigma_Asp}, \code{Ae_Glu}, \code{Ea_Glu}, \code{x_Glu}, \code{sigma_Glu}, \code{rho}.
}
\description{
Extract posterior-mean forward-model calibration parameters
}
