% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APDImem.R
\name{NME}
\alias{NME}
\title{Assessing the Performance of Dietary Assessment Instruments}
\usage{
NME(Q.N.1, Q.N.2, Q.E.1, Q.E.2, F.N.1, F.N.2, F.E.1, F.E.2)
}
\arguments{
\item{data}{data has to be complete data.}

\item{Q.N.1:}{An assessment instrument for nutrient (N) at time 1.}

\item{Q.N.2:}{An assessment instrument for nutrient (N) at time 2.}

\item{Q.E.1:}{An assessment instrument for total energy (E) at time 1.}

\item{Q.E.2:}{An assessment instrument for total energy (E) at time 2.}

\item{F.N.1:}{A reference instrument for nutrient (N) at time 1.}

\item{F.N.2:}{A reference instrument for nutrient (N) at time 2.}

\item{F.E.1:}{A reference instrument for total energy (E) at time 1.}

\item{F.E.2:}{A reference instrument for total energy (E) at time 2.}
}
\value{
\item{Rho.N}{correlation coefficient between reported and true intake of nutrient N.}
\item{Rho.E}{correlation coefficient between reported and true intake of total energy E.}
\item{Rho.R}{correlation coefficient between reported and true intake residuals.}
\item{Lambda.N}{attenuation factor of nutrient N. }
\item{Lambda.E}{attenuation factor of total energy E.}
\item{Lambda.R}{attenuation factor of nutrient residual R.}
\item{est.covariance.matrix}{estimated covariance matrix for observed data}
\item{Beta.N1}{intake-related bias of nutrient N in the FFQ}
\item{Beta.E1}{intake-related bias of total energy E in the FFQ}
\item{Mu.T.N}{mean of ture intake of nutrient N}
\item{Mu.T.E}{mean of ture intake of total energy E}
\item{Mu.N0.2}{overall bias independent of true intake of nutrient N for the 24HR}
\item{Mu.E0.2}{overall bias independent of true intake of total energy E for the 24HR}
\item{Beta.N0.1}{overall bias independent of true intake of nutrient N for the FFQ in the first measurement}
\item{Beta.N0.2}{overall bias independent of true intake of total energy E for the FFQ in the first measurement}
\item{Beta.E0.1}{overall bias independent of true intake of nutrient N for the FFQ in the first measurement}
\item{Beta.E0.2}{overall bias independent of true intake of total energy E for the FFQ in the first measurement}
}
\description{
Assessing the Performance of Dietary Assessment Instruments
}
\details{
A package for estimating both of the correlations of a dietary assessment instrument (Q) and nutrient intake and attenuation factors and the correlations of nutrient residual.
The model is established as the measurement error model to evaluate the data collected by a food-frequency questionnaire (FFQ) using two non-consecutive 24-hour dietary recalls (24HRs) as the reference.
We employ the moment estimation method to the complete longitudinal data structure. Moreover, our method can provide the estimates of fixed effects including intake-related biases in the FFQ and overall biases of true intake for the FFQ and 24HR, respectively in the model.
Notably, the correlations with true intake and attenuation factors for energy-adjusted should be interpreted on the scale to which each particular nutrient was transformed.
}
\examples{
## example using random numbers

library(MASS)
set.seed(1)
# number of observed individuals:
I<- 100
# fixed effects in the measurement error model:
# beta = (beta_N0_1, beta_N0_2, beta_N1, beta_E0_1, beta_E0_2, beta_E1)
beta<-  c(0.4, 0.5, 0.3,0.32,0.3,0.35)
# mu_1 = (mu_T_N, mu_T_E, mu_N0_2, mu_E0_2)
mu_1<-  c(0.3,0.32,0.4,0.35)
# mu = first moments:  the mean vector of D_i
mu<- rep(0,8)
mu[1]<- beta[1] + beta[3]*mu_1[1]
mu[2]<- beta[2] + beta[3]*mu_1[1]
mu[3]<- beta[4] + beta[6]*mu_1[2]
mu[4]<- beta[5] + beta[6]*mu_1[2]
mu[5]<- mu_1[1]
mu[6]<- mu_1[3] + mu_1[1]
mu[7]<- mu_1[2]
mu[8]<- mu_1[4] + mu_1[2]
# second moments:
# theta = (theta1, theta2, ..., theta15)
# we notice that theta1, theta7, theta11, theta12, theta14, theta15 are variance components, then they should be set as positive numbers.
theta<- rep(0,15)
# the setting of theta:
theta<- c(0.5, 0, 0, 0, 0.175 , 0, 0.58, 0, 0, 0.232, 0.5, 0.35, 0 , 0.55 ,0.3)
theta_set<- c(theta[1], theta[2], theta[3], theta[4], theta[5], theta[5], theta[6], theta[6], theta[2], theta[1], theta[4], theta[3], theta[5], theta[5], theta[6], theta[6],
             theta[3], theta[4], theta[7], theta[8], theta[9], theta[9], theta[10], theta[10], theta[4], theta[3], theta[8], theta[7], theta[9], theta[9], theta[10], theta[10],
             theta[5], theta[5], theta[9], theta[9], theta[11], theta[12], theta[13],theta[13],
             theta[5], theta[5], theta[9], theta[9], theta[12], theta[11], theta[13],theta[13],
             theta[6], theta[6], theta[10], theta[10], theta[13], theta[13], theta[14], theta[15], theta[6], theta[6],theta[10], theta[10], theta[13], theta[13], theta[15], theta[14])
# variance covariance matrix of D_i
# the observed data:
D<- matrix(0, nrow=8 , ncol=I)
D<- t(as.matrix(mvrnorm(I, mu, Sigma)))
Data<- data.frame(t(D))
colnames(Data)<- c("FFQ.N.1","FFQ.N.2", "FFQ.E.1", "FFQ.E.2", "Ref.N.1", "Ref.N.2","Ref.E.1", "Ref.E.2")
ex1<- NME(Q.N.1=Data$FFQ.N.1, Q.N.2=Data$FFQ.N.2, Q.E.1= Data$FFQ.E.1, Q.E.2= Data$FFQ.E.2, F.N.1= Data$Ref.N.1, F.N.2= Data$Ref.N.2, F.E.1= Data$Ref.E.1, F.E.2= Data$Ref.E.2)
print(ex1)

}
\references{
Thompson, F. E., Kipnis, V., Midthune, D., Freedman, L.S., Carroll, R.J., Subar, A.F., Brown, C.C., Butcher, M.S.,  Mouw, T., Leitzmann, M. and Schatzkin, A. (2008) Performance of a food-frequency questionnaire in the us NIH--AARP (National Institutes of Health--American Association of Retired Persons) Diet and Health Study. \emph{Public Health Nutrition}, 11, 183-195.
}
